rename managed pool
